Sync 2.3.x build with master, which includes the travis build optimization.
