+++ .travis.yml
@@ -1,7 +1,7 @@
 language: scala
 scala:
+  - 2.11.5
   - 2.10.4
-  - 2.11.2
 jdk:
   - oraclejdk7
   - openjdk6
@@ -16,20 +16,35 @@ env:
     - SCALATEST_GPG_PASSPHASE=tbd
 
   matrix:
-    - MODE=Main
-    - MODE=genMustMatchersTests
+    - MODE=RegularTests1
+    - MODE=RegularTests2
+    - MODE=RegularTests3
+    - MODE=RegularTests4
+    - MODE=RegularTests5
+    - MODE=ScalacticTests
+    - MODE=genMustMatchersTests1
+    - MODE=genMustMatchersTests2
+    - MODE=genMustMatchersTests3
+    - MODE=genMustMatchersTests4
     - MODE=genGenTests
     - MODE=genTablesTests
     - MODE=genInspectorsTests
-    - MODE=genInspectorsShorthandsTests
+    - MODE=genInspectorsShorthandsTests1
+    - MODE=genInspectorsShorthandsTests2
     - MODE=genTheyTests
-    - MODE=genContainTests
+    - MODE=genContainTests1
+    - MODE=genContainTests2
     - MODE=genSortedTests
     - MODE=genLoneElementTests
     - MODE=genEmptyTests
 
 #before_script: ./travis_build.sh Compile
 script: ./travis_build.sh $MODE
+
+notifications:
+  email:
+    - bill@artima.com
+    - cheeseng@amaseng.com
 
 #after_success:
 #  # only 'scalatest/scalatest' 'master' branch is published from the first node
